Move operation route factory
